* Killed a big leak in the paginator. * Killed another warning.
git-svn-id: https://svn.xiph.org/trunk/oggdsf@8715 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/tests/testLibOOOgg/TestPatternOggPageChecker.cpp
@@ -12,7 +12,7 @@ TestPatternOggPageChecker::~TestPatternOggPageChecker(void)
 bool TestPatternOggPageChecker::acceptOggPage(OggPage* inOggPage) 
 {
 	bool locIsOK = true;
-	for (int i = 0; i < inOggPage->numPackets(); i++) {
+	for (unsigned long i = 0; i < inOggPage->numPackets(); i++) {
 		locIsOK = locIsOK && (mPacketChecker.acceptStampedOggPacket( (StampedOggPacket*)(inOggPage->getStampedPacket(i)->clone()) ) );
 	}
